next: move the features/openai-chatgpt page to ./ai-llm as a start to update it and make it more generic
